requirements.txt
bson==0.5.8
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.1
dnspython==1.16.0
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.12.3
idna==2.8
psycopg2==2.8.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pymongo==3.8.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
ttldict==0.3.0
uritemplate==3.0.0
urllib3==1.25.2
whitenoise==4.1.2
